LICENSE
Extended Description
Licensing Information

Copyright (c) 2025 Ismail

This project is licensed under a Proprietary License. By using this software, you agree to the following terms and conditions:
Usage Rights

    You are granted a personal, non-exclusive, and non-transferable right to use this software.
    This software may only be used for personal, non-commercial purposes unless otherwise agreed upon with the author.
    Redistribution or sharing of this software, in any form, is strictly prohibited without prior written consent from the author.

Modification

    You are not allowed to modify, adapt, or otherwise alter the software in any way without prior written consent from the author.
    If you wish to extend or modify the software, you must obtain approval from the author before redistributing any modified versions.

Commercial Use

    Commercial use of this software is strictly prohibited unless explicitly agreed upon with the author. If you intend to use the software for commercial purposes, you must reach out to the author for a commercial licensing agreement.

Warranty Disclaimer

This software is provided "AS IS," without any warranties or conditions of any kind. The author does not guarantee that the software will be free from errors or suitable for any particular purpose.
Limitation of Liability

In no event shall the author or contributors be liable for any damages, including but not limited to direct, indirect, incidental, special, or consequential damages, arising from the use or inability to use the software.